pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "sd-order-gui"
version = "0.1.0"
description = "Stellar Dominion PBEM turn importer and order YAML generator (GUI)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.6",
  "PyYAML>=6.0",
]

[project.scripts]
sd-order-gui = "sd_order_gui.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["sd_order_gui"]

